Changed use of PublishSubject to ReplaySubject for consistency with WebSockets behaviour.

+++ src/main/java/io/github/ngchinhow/kraken/websocket/model/message/ohlc/WebSocketsOHLC.java
@@ -30,7 +30,7 @@ public class WebSocketsOHLC {
     @NonNull
     private BigDecimal vwap;
     @NonNull
-    private BigDecimal interval;
+    private Integer interval;
     @NonNull
     private ZonedDateTime timestamp;
 }
